Update eodrcpt.hsl
I've updated the code to use $transaction["id"] instead of $messageid. $messageid isn't available in eodrcpt so the script failed on 'check code'. 

Several of the scripts are still using $recepient and $sender, might want to look at that. post.hsl is also using $messageid, but the variable is available.

https://support.halon.io/hc/en-us/articles/360002469740
- Search for 'Renamed variables'